+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DashApp.get_app_layout: added `className="dbc"` to support Bootstrap CSS, see https://hellodash.pythonanywhere.com/about_dbc_css
 - DashApp and DoDashApp constructors: added bootstrap_theme and bootstrap_figure_template to support Bootstrap CSS
 - PlotlyRowsVisualizationPage template to easily add a visualization page of rows of Plotly Figures
+- utils.domodelrunner and utils.donotebookrunner modules
+- RunModel page with IO for running in-line and long-running model callbacks
 
 ## [0.1.1.0] - 2022-01-11
 ### Changed